+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   * Interactive Orientation arrow not moving with azimuth picker arrow is moved.
   * Azimuth picker arrow not moving when interactive orientation arrow is moved.
   * LSize text value in lsize option window does not change when lsize is interactivelly changed.
+  * LSize interactive arrow size does not change when LSize text value in lsize option window changes.
 * Infrastructure maintenance:
   * Testing all possible variations of scrap scale option parsing.
   * Custom exceptions should provide trace.